Add Dockerfile and docker-compose, update .gitignore and set_config script
This commit adds Dockerfile and docker-compose files to prepare the application for containerization. It also modifies the .gitignore file changing the config file extension from '.yaml' to '.yml'. An extended 'set_config.sh' script is included for generating random service API keys and adding them to a .env file if they don't exist. These changes are key for automating and enhancing the deployment process.