Update README.md with service running instructions
Added detailed information on how to run the service using docker. The newly included instructions guide through how to copy the environment settings and build the relevant containers. Additional commands for starting and ending the service are also listed. This improves user understanding and simplifies service operation for the users.